guarda pedidos y productos

+++ app/src/main/java/com/example/jaime/homeserviceoficial/Database/DataSource/CartRepository.java
@@ -39,7 +39,7 @@ public class CartRepository implements ICartDataSource{
     }
 
     @Override
-    public float sumPrecio() {
+    public double sumPrecio() {
         return iCartDataSource.sumPrecio();
     }
