+++ app/src/main/java/com/example/jaime/homeserviceoficial/Database/Local/CartDataSource.java
@@ -38,7 +38,7 @@ public class CartDataSource implements ICartDataSource {
     }
 
     @Override
-    public float sumPrecio() {
+    public double sumPrecio() {
         return cartDAO.sumPrecio();
     }
